flask_app/requirements.txt
opencv-python==4.8.1.78
numpy==1.26.1
pandas==2.1.3
mtcnn==0.1.1
#
tensorflow==2.14.0
keras-facenet==0.3.2
scikit-learn==1.3.2
scipy==1.11.3
Pillow==10.1.0
#
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.0.1
Flask-Login==0.6.3
